Refactor JWT token validation tests and update issuer URLs
- Updated issuer URLs in JwtTokenValidatorTest to reflect new domain.
- Added tests for optional email and organization claims in JWT.
- Ensured audience array is accepted and issued at claim is required.
- Refactored ProductionAuthenticationContextTest to use InMemoryExternalIdentityRepository.
- Added tests for handling suspended external identities and ambiguous organization memberships.
- Updated organization-related tests to reflect new naming conventions.
- Introduced InMemoryExternalIdentityRepository for better test isolation.
- Removed deprecated methods from InMemoryUserRepository.
- Adjusted LocalImportFileStorageTest to use new directory naming.
- Removed obsolete IndustrialFoundationTest.
- Updated database identifier mappings to use bigint.
- Added PlatformFoundationBaselineTest to validate migration structure.
- Enhanced FoundationUseCasesTest and OwnerProtectionTest with audit logging.
- Created IdentityOrganizationAccessTest to validate organization access logic.
- Updated OpenApiDocumentationTest to reflect new API paths and security schemes.

+++ tests/Infrastructure/Doctrine/IntegerIdentifierMappingTest.php
@@ -19,7 +19,7 @@ final class IntegerIdentifierMappingTest extends KernelTestCase
         self::assertNotEmpty($all);
         foreach ($all as $mapping) {
             self::assertSame(['id'], $mapping->getIdentifierFieldNames(), $mapping->name);
-            self::assertSame('integer', $mapping->getTypeOfField('id'), $mapping->name);
+            self::assertSame('bigint', $mapping->getTypeOfField('id'), $mapping->name);
             self::assertSame(ClassMetadata::GENERATOR_TYPE_IDENTITY, $mapping->generatorType, $mapping->name);
             self::assertTrue($mapping->getFieldMapping('id')->options['unsigned'] ?? false, $mapping->name);
         }
